// src/app/models/vehiculo.model.ts
export interface Vehiculo {
    placa: string;
    numeroMotor: string;
    numeroChasis: string;
    VIN: string;
    marca: string;
    clase: string;
    linea: string;
    modelo: string;
    color: string;
    cilindraje: string;
    potencia: string;
    capacidad: string;
    servicio: string;
    carroceria: string;
    combustible: string;
}